sca: fix compile by moving a few things around
